# app/public/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "This api full project in Rosinfra sercvice",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8300/api/",
            "description": "The baerer security filter is used for authorization on the service. You can get the token in your merchant profile."
        }
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\User\\ProfileController::show",
                "requestBody": {
                    "description": "order placed for purchasing th pet",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileResource"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check your email and click on the link to verify.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProfileResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\User\\ProfileController::update",
                "requestBody": {
                    "description": "order placed for purchasing th pet",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileEditRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check your email and click on the link to verify.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProfileEditRequest"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": [],
                        "OAuth2": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ProfileEditRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ProfileResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "properties": {
                            "number": {
                                "type": "string"
                            },
                            "verified": {
                                "type": "boolean"
                            }
                        },
                        "type": "object"
                    },
                    "name": {
                        "properties": {
                            "first": {
                                "type": "string"
                            },
                            "last": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "oauth2",
                "in": "header",
                "flows": {
                    "implicit": {
                        "tokenUrl": "https://localhost:8300/oauth/token",
                        "scopes": {
                            "write:adverts": "modify pets in your account",
                            "read:adverts": "read your pets"
                        }
                    }
                }
            }
        }
    }
}